requirements.txt
flask==3.0.3
gunicorn==22.0.0
dnspython==2.6.1
pytest==8.2.2
pytest-cov==5.0.0
